+++ src/main/java/com/docente/proyectoejemplo/mascotas/MascotasRepo.java
@@ -7,6 +7,6 @@ import com.docente.proyectoejemplo.mascotas.entity.Mascota;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 @Repository 
-public interface MascotasRepo extends JpaRepository<Mascota, Short>{
+public interface MascotasRepo extends JpaRepository<Mascota, Integer>{
     
 }
